backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg[binary]
python-dotenv
pydantic-settings
argon2-cffi
python-jose[cryptography]
alembic
email-validator
